+++ python/weka/associations.py
@@ -12,7 +12,7 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 # associations.py
-# Copyright (C) 2014-2016 Fracpete (pythonwekawrapper at gmail dot com)
+# Copyright (C) 2014-2018 Fracpete (pythonwekawrapper at gmail dot com)
 
 import javabridge
 import logging
@@ -21,11 +21,492 @@ import sys
 import argparse
 import weka.core.jvm as jvm
 import weka.core.converters as converters
-from weka.core.classes import OptionHandler, join_options
+from weka.core.classes import OptionHandler, JavaObject, join_options
 from weka.core.capabilities import Capabilities
+from weka.core.dataset import Attribute
+from weka.core.typeconv import string_array_to_list
 
 # logging setup
 logger = logging.getLogger("weka.associations")
+
+
+class Item(JavaObject):
+    """
+    Wrapper for weka.associations.Item class.
+    """
+
+    def __init__(self, jobject):
+        """
+        Initializes the wrapper.
+
+        :param jobject: the Item object to wrap
+        :type jobject: JB_Object
+        """
+
+        if jobject is None:
+            raise Exception("No Item JB_Object provided!")
+        self.enforce_type(jobject, "weka.associations.Item")
+        super(Item, self).__init__(jobject)
+
+    def __compareto__(self, other):
+        """
+        Return the comparison result with the other rule.
+
+        :param other: the other rule to compare with
+        :type other: Item
+        :return: the integer result, <0 is less than, =0 is equal to, >0 is greater than
+        :rtype: int
+        """
+        return javabridge.call(self.jobject, "compareTo", "(Lweka/associations/Item;)I", other.jobject)
+
+    def __lt__(self, other):
+        """
+        Checks self < other.
+
+        :param other: the item to compare with
+        :type other: Item
+        :return: whether less than the other item
+        :rtype: bool
+        """
+        return self.__compareto__(other) < 0
+
+    def __le__(self, other):
+        """
+        Checks self <= other.
+
+        :param other: the item to compare with
+        :type other: Item
+        :return: whether less or equal than the other item
+        :rtype: bool
+        """
+        return self.__compareto__(other) <= 0
+
+    def decrease_frequency(self, frequency=None):
+        """
+        Decreases the frequency.
+
+        :param frequency: the frequency to decrease by, 1 if None
+        :type frequency: int
+        """
+        if frequency is None:
+            javabridge.call(self.jobject, "decreaseFrequency", "()V")
+        else:
+            javabridge.call(self.jobject, "decreaseFrequency", "(I)V", frequency)
+
+    def increase_frequency(self, frequency=None):
+        """
+        Increases the frequency.
+
+        :param frequency: the frequency to increase by, 1 if None
+        :type frequency: int
+        """
+        if frequency is None:
+            javabridge.call(self.jobject, "increaseFrequency", "()V")
+        else:
+            javabridge.call(self.jobject, "increaseFrequency", "(I)V", frequency)
+
+    def __eq__(self, other):
+        """
+        Returns whether this object is the same as the other one.
+
+        :param other: the object to compare with
+        :type other: Item
+        :return: True if the same
+        :rtype: bool
+        """
+        return javabridge.call(self.jobject, "equals", "(Ljava/lang/Object;)Z", other.jobject)
+
+    def __ge__(self, other):
+        """
+        Checks self >= other.
+
+        :param other: the item to compare with
+        :type other: Item
+        :return: whether greater or equal than the other item
+        :rtype: bool
+        """
+        return self.__compareto__(other) >= 0
+
+    def __gt__(self, other):
+        """
+        Checks self > other.
+
+        :param other: the item to compare with
+        :type other: Item
+        :return: whether greater than the other item
+        :rtype: bool
+        """
+        return self.__compareto__(other) > 0
+
+    def __str__(self):
+        """
+        Returns the item as string.
+
+        :return: the item as string
+        :rtype: str
+        """
+        return javabridge.call(self.jobject, "toString", "()Ljava/lang/String;")
+
+    @property
+    def frequency(self):
+        """
+        Returns the frequency.
+
+        :return: the frequency
+        :rtype: int
+        """
+        return javabridge.call(self.jobject, "getFrequency", "()I")
+
+    @property
+    def attribute(self):
+        """
+        Returns the attribute.
+
+        :return: the attribute
+        :rtype: Attribute
+        """
+        return Attribute(javabridge.call(self.jobject, "getAttribute", "()Lweka/core/Attribute;"))
+
+    @property
+    def comparison(self):
+        """
+        Returns the comparison operator as string.
+
+        :return: the comparison iterator
+        :rtype: str
+        """
+        return Attribute(javabridge.call(self.jobject, "getComparisonAsString", "()Ljava/lang/String;"))
+
+    @property
+    def item_value(self):
+        """
+        Returns the item value as string.
+
+        :return: the item value
+        :rtype: str
+        """
+        return Attribute(javabridge.call(self.jobject, "getItemValueAsString", "()Ljava/lang/String;"))
+
+
+class AssociationRule(JavaObject):
+    """
+    Wrapper for weka.associations.AssociationRule class.
+    """
+
+    def __init__(self, jobject):
+        """
+        Initializes the wrapper.
+
+        :param jobject: the AssociationRule object to wrap
+        :type jobject: JB_Object
+        """
+        if jobject is None:
+            raise Exception("No AssociationRule JB_Object provided!")
+        self.enforce_type(jobject, "weka.associations.AssociationRule")
+        super(AssociationRule, self).__init__(jobject)
+
+    def __compareto__(self, other):
+        """
+        Return the comparison result with the other rule.
+
+        :param other: the other rule to compare with
+        :type other: AssociationRule
+        :return: the integer result, <0 is less than, =0 is equal to, >0 is greater than
+        :rtype: int
+        """
+        return javabridge.call(self.jobject, "compareTo", "(Lweka/associations/AssociationRule;)I", other.jobject)
+
+    def __lt__(self, other):
+        """
+        Checks self < other.
+
+        :param other: the rule to compare with
+        :type other: AssociationRule
+        :return: whether less than the other rule
+        :rtype: bool
+        """
+        return self.__compareto__(other) < 0
+
+    def __le__(self, other):
+        """
+        Checks self <= other.
+
+        :param other: the rule to compare with
+        :type other: AssociationRule
+        :return: whether less or equal than the other rule
+        :rtype: bool
+        """
+        return self.__compareto__(other) <= 0
+
+    def __eq__(self, other):
+        """
+        Returns whether this object is the same as the other one.
+
+        :param other: the object to compare with
+        :type other: AssociationRule
+        :return: True if the same
+        :rtype: bool
+        """
+        return javabridge.call(self.jobject, "equals", "(Ljava/lang/Object;)Z", other.jobject)
+
+    def __ge__(self, other):
+        """
+        Checks self >= other.
+
+        :param other: the rule to compare with
+        :type other: AssociationRule
+        :return: whether greater or equal than the other rule
+        :rtype: bool
+        """
+        return self.__compareto__(other) >= 0
+
+    def __gt__(self, other):
+        """
+        Checks self > other.
+
+        :param other: the rule to compare with
+        :type other: AssociationRule
+        :return: whether greater than the other rule
+        :rtype: bool
+        """
+        return self.__compareto__(other) > 0
+
+    @property
+    def consequence(self):
+        """
+        Get the the consequence.
+
+        :return: the consequence, list of Item objects
+        :rtype: list
+        """
+        items = javabridge.get_collection_wrapper(
+            javabridge.call(self.jobject, "getConsequence", "()Ljava/util/Collection;"))
+        result = []
+        for item in items:
+            result.append(Item(item))
+        return result
+
+    @property
+    def consequence_support(self):
+        """
+        Get the support for the consequence.
+
+        :return: the support
+        :rtype: int
+        """
+        return javabridge.call(self.jobject, "getConsequenceSupport", "()I")
+
+    @property
+    def premise(self):
+        """
+        Get the the premise.
+
+        :return: the premise, list of Item objects
+        :rtype: list
+        """
+        items = javabridge.get_collection_wrapper(
+            javabridge.call(self.jobject, "getPremise", "()Ljava/util/Collection;"))
+        result = []
+        for item in items:
+            result.append(Item(item))
+        return result
+
+    @property
+    def premise_support(self):
+        """
+        Get the support for the premise.
+
+        :return: the support
+        :rtype: int
+        """
+        return javabridge.call(self.jobject, "getPremiseSupport", "()I")
+
+    @property
+    def total_support(self):
+        """
+        Get the total support.
+
+        :return: the support
+        :rtype: int
+        """
+        return javabridge.call(self.jobject, "getTotalSupport", "()I")
+
+    @property
+    def total_transactions(self):
+        """
+        Get the total transactions.
+
+        :return: the transactions
+        :rtype: int
+        """
+        return javabridge.call(self.jobject, "getTotalTransactions", "()I")
+
+    @property
+    def metric_names(self):
+        """
+        Returns the metric names for the rule.
+
+        :return: the metric names
+        :rtype: list
+        """
+        return string_array_to_list(javabridge.call(self.jobject, "getMetricNamesForRule", "()[Ljava/lang/String;"))
+
+    @property
+    def metric_values(self):
+        """
+        Returns the metric values for the rule.
+
+        :return: the metric values
+        :rtype: ndarray
+        """
+        return javabridge.get_env().get_double_array_elements(
+            javabridge.call(self.jobject, "getMetricValuesForRule", "()[D"))
+
+    def metric_value(self, name):
+        """
+        Returns the named metric value for the rule.
+
+        :param name: the name of the metric
+        :type name: str
+        :return: the metric value
+        :rtype: float
+        """
+        return javabridge.call(self.jobject, "getNamedMetricValue", "(Ljava/lang/String;)D", name)
+
+    @property
+    def primary_metric_name(self):
+        """
+        Returns the primary metric name for the rule.
+
+        :return: the metric name
+        :rtype: str
+        """
+        return javabridge.call(self.jobject, "getPrimaryMetricName", "()Ljava/lang/String;")
+
+    @property
+    def primary_metric_value(self):
+        """
+        Returns the primary metric value for the rule.
+
+        :return: the metric value
+        :rtype: float
+        """
+        return javabridge.call(self.jobject, "getPrimaryMetricValue", "()D")
+
+
+class AssociationRulesIterator(object):
+    """
+    Iterator for weka.associations.AssociationRules class.
+    """
+
+    def __init__(self, rules):
+        """
+        Initializes with the rules.
+
+        :param rules: the rules to use
+        :type rules: AssociationRules
+        """
+        self.rules = rules
+        self.index = 0
+        self.length = len(rules)
+
+    def __iter__(self):
+        """
+        Returns itself.
+        """
+        return self
+
+    def __next__(self):
+        """
+        Returns the next rule.
+
+        :return: the next rule object
+        :rtype: AssociationRule
+        """
+        if self.index < self.length:
+            index = self.index
+            self.index += 1
+            return self.rules[index]
+        else:
+            raise StopIteration()
+
+
+class AssociationRules(JavaObject):
+    """
+    Wrapper for weka.associations.AssociationRules class.
+    """
+
+    def __init__(self, jobject):
+        """
+        Initializes the wrapper.
+
+        :param jobject: the AssociationRules object to wrap
+        :type jobject: JB_Object
+        """
+        if jobject is None:
+            raise Exception("No AssociationRules JB_Object provided!")
+        self.enforce_type(jobject, "weka.associations.AssociationRules")
+        super(AssociationRules, self).__init__(jobject)
+
+    def __len__(self):
+        """
+        Returns the number of rules available.
+
+        :return: the number of rules
+        :rtype: int
+        """
+        return javabridge.call(self.jobject, "getNumRules", "()I")
+
+    def __getitem__(self, item):
+        """
+        Returns the specified item, using 0-based indices.
+
+        :param item: the 0-based index
+        :type item: int
+        :return: the AssociationRule
+        :rtype: AssociationRule
+        """
+        rules = javabridge.get_collection_wrapper(
+            javabridge.call(self.jobject, "getRules", "()Ljava/util/List;"))
+        return AssociationRule(rules[item])
+
+    def __str__(self):
+        """
+        Returns a short description of the rules.
+
+        :return: short description
+        :rtype: str
+        """
+        return str(len(self)) + " rules generated by " + self.producer
+
+    def __iter__(self):
+        """
+        Returns an iterator for the rules.
+
+        :return: the iterator
+        :rtype: AssociationRulesIterator
+        """
+        return AssociationRulesIterator(self)
+
+    @property
+    def producer(self):
+        """
+        Returns a string describing the producer that generated these rules.
+
+        :return: the producer
+        :rtype: str
+        """
+        return javabridge.call(self.jobject, "getProducer", "()Ljava/lang/String;")
+
+    @producer.setter
+    def producer(self, producer):
+        """"
+        Sets the string describing the producer that generated these rules.
+
+        :param producer: the producer
+        :type producer: str
+        """
+        javabridge.call(self.jobject, "setProducer", "(Ljava/lang/String;)V", producer)
 
 
 class Associator(OptionHandler):
@@ -67,6 +548,43 @@ class Associator(OptionHandler):
         :type data: Instances
         """
         javabridge.call(self.jobject, "buildAssociations", "(Lweka/core/Instances;)V", data.jobject)
+
+    def can_produce_rules(self):
+        """
+        Checks whether association rules can be generated.
+
+        :return: whether scheme implements AssociationRulesProducer interface and
+        association rules can be generated
+        :rtype: bool
+        """
+        if not self.check_type(self.jobject, "weka.associations.AssociationRulesProducer"):
+            return False
+        return javabridge.call(self.jobject, "canProduceRules", "()Z")
+
+    def association_rules(self):
+        """
+        Returns association rules that were generated. Only if implements AssociationRulesProducer.
+
+        :return: the association rules that were generated
+        :rtype: AssociationRules
+        """
+        if not self.check_type(self.jobject, "weka.associations.AssociationRulesProducer"):
+            return None
+        return AssociationRules(
+            javabridge.call(self.jobject, "getAssociationRules", "()Lweka/associations/AssociationRules;"))
+
+    @property
+    def rule_metric_names(self):
+        """
+        Returns the rule metric names of the association rules. Only if implements AssociationRulesProducer.
+
+        :return: the metric names
+        :rtype: list
+        """
+        if not self.check_type(self.jobject, "weka.associations.AssociationRulesProducer"):
+            return None
+        return string_array_to_list(
+            javabridge.call(self.jobject, "getRuleMetricNames", "()[Ljava/lang/String;"))
 
     @classmethod
     def make_copy(cls, associator):